Allow git group read-only access to git home dir
Everything but .ssh below the home dir was already 755 perms,
but the homedir was preventing users added to the 'git' group
from actually having read access to the repositories. Group
read access is a useful thing with integration to development apps
(Trac, Redmine, ReviewBoard, etc) so go ahead and enable it.
The ~/.ssh dir of course keeps the stricter owner-only access.